[Refactor] Service 인터페이스화
